test: fzf_lua integration
